# microservices/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample swagger docs for argus project",
        "title": "Argus API docs",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "fiber@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/alarm": {
            "post": {
                "description": "Set car to security",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Set car to security",
                "operationId": "alarm-on",
                "parameters": [
                    {
                        "description": "AlarmOnRequest",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AlarmOnRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AlarmOnResponse"
                        }
                    },
                    "400": {
                        "description": "Struct is invalid",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Cannot request message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AlarmOnRequest": {
            "type": "object",
            "required": [
                "car_id",
                "login",
                "time"
            ],
            "properties": {
                "car_id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "models.AlarmOnResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        }
    }
}